refactored Artist and Song classes with Memorable/Findable/Paramable modules (also used super)
