+++ lib/concerns/findable.rb
@@ -1,0 +1,5 @@
+module Findable
+  def find_by_name(name)
+    self.all.detect{|a| a.name == name}
+  end
+end
