+++ lib/concerns/paramable.rb
@@ -1,0 +1,5 @@
+module Paramable
+  def to_param
+    name.downcase.gsub(' ', '-')
+  end
+end
