+++ lib/song.rb
@@ -3,34 +3,19 @@ require 'pry'
 class Song
   attr_accessor :name
   attr_reader :artist
-
+  extend Memorable::ClassMethods, Findable
+  include Paramable, Memorable::InstanceMethods
+  
   @@songs = []
 
-  def initialize
-    @@songs << self
-  end
 
-  def self.find_by_name(name)
-    @@songs.detect{|a| a.name == name}
-  end
 
   def self.all
     @@songs
-  end
-
-  def self.reset_all
-    self.all.clear
-  end
-
-  def self.count
-    self.all.count
   end
 
   def artist=(artist)
     @artist = artist
   end
 
-  def to_param
-    name.downcase.gsub(' ', '-')
-  end
 end
